backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.92.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.135.3
google-ai-generativelanguage==0.6.15
google-analytics-admin==0.28.0
google-analytics-data==0.21.0
google-api-core==2.30.2
google-api-python-client==2.194.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-genai==1.72.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.27.0
oauthlib==3.3.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pyjwt==2.12.1
pyparsing==3.3.2
python-dotenv==1.2.2
python-multipart==0.0.24
pyyaml==6.0.3
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rpds-py==0.30.0
sniffio==1.3.1
sse-starlette==3.3.4
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
